Move LsEtcSsh from ls_etc_ssh.py to ls_etc.py
