--- docs/shared_parsers_catalog/ls_etc_ssh.rst
@@ -1,3 +1,0 @@
-.. automodule:: insights.parsers.ls_etc_ssh
-   :members:
-   :show-inheritance:
